add some comment

+++ src/log/block_queue.h
@@ -11,8 +11,8 @@ class BlockQueue {
   explicit BlockQueue(size_t cap = 1000);
   bool Push(const T &item);
   bool Pop(T &item);
-  bool Pop(T &item, int ms_timeout);
-  void Close();
+  bool Pop(T &item, size_t ms_timeout);  // 设置超时时间
+  void Close();                          // 关闭阻塞队列
  private:
   bool IsFull();
   bool IsEmpty();
